requirements.txt
weave==0.50.7
wandb>=0.16.0
openai>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
asyncio
aiohttp>=3.8.0
pydantic>=2.0.0
tiktoken>=0.5.0
openpipe-ai  # Optional - for RL training data collection
flask>=2.0.0
streamlit>=1.28.0
plotly>=5.0.0